--- sdk/storage/Azure.Storage.Blobs/tests/SessionRecords/BlobBaseClientTests/BlobDownloadInfo_Dispose.json
@@ -1,4 +1,0 @@
-{
-  "Entries": [],
-  "Variables": {}
-}
