--- sdk/storage/Azure.Storage.Blobs/tests/SessionRecords/BlobBaseClientTests/Disposable.json
@@ -1,4 +1,0 @@
-{
-  "Entries": [],
-  "Variables": {}
-}
