+++ sdk/storage/Azure.Storage.Files/src/Models/StorageFileDownloadInfo.cs
@@ -53,7 +53,7 @@ namespace Azure.Storage.Files.Models
         /// <summary>
         /// Disposes the StorageFileDownloadInfo by calling Dispose on the underlying Content stream.
         /// </summary>
-        public void Dispose()
+        public virtual void Dispose()
         {
             Content?.Dispose();
             GC.SuppressFinalize(this);
