--- sdk/storage/Azure.Storage.Files/tests/SessionRecords/FileClientTests/StorageFileDownloadInfo_Dispose.json
@@ -1,4 +1,0 @@
-{
-  "Entries": [],
-  "Variables": {}
-}
